move block creation out of fast forward
