+++ core/Cargo.toml
@@ -30,6 +30,7 @@ zstd = { workspace = true }
 
 frame-remote-externalities = { workspace = true }
 frame-try-runtime = { workspace = true, features=["try-runtime"] }
+frame-support = { workspace = true }
 
 sc-cli = { workspace = true }
 sc-executor = { workspace = true }
