--- core/src/common/empty_block_creation/mod.rs
@@ -1,1 +1,0 @@
-pub mod inherents;
