+++ core/src/common/mod.rs
@@ -1,4 +1,4 @@
-pub mod empty_block_creation;
+pub mod empty_block;
 pub mod misc_logging;
 pub mod parse;
 pub mod shared_parameters;
